--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,11 +6,19 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.70"
+async-std = "1.12.0"
 bitcoin = { version = "0.30.0" }
 bitflags = "2.0.1"
 byteorder = "1.4.3"
 more-asserts = "0.3.1"
+mp4 = "0.13.0"
+ord-labs = { version = "0.1.0", git = "https://github.com/ordilabs/ord-labs.git", branch = "main" }
+postgres = "0.19.5"
 rusty-leveldb = "1.0.6"
+serde = "1.0.159"
+serde_derive = "1.0.159"
+serde_json = "1.0.95"
 
 [dev-dependencies]
 directories = "5.0.0"