significant speedup by using `BufReader` to lower # syscalls

testrun with 1000 blocks:
```bash
# before
cargo run -r --example simple_mapping  23.97s user 60.00s system 78% cpu 1:47.14 total
# after
cargo run -r --example simple_mapping  5.93s user 3.32s system 46% cpu 20.058 total
```